Use postinstall script to download bin!
