--- bin/android-x64/ssb-patchql
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:878b50861f5092fffc2357a1a4733ebdda73468f62c1e748231526c1ec9466bb
-size 7261968
