--- bin/darwin-x64/ssb-patchql
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:b99eb942e01c6fcabf3b401b4760cccc58a98cc2c6a559fd497ffb18994180f2
-size 7610052
