--- bin/linux-arm64/ssb-patchql
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:2c81b93f350c86aea58c7e375f94180d826ae8c8e36d051b65bdf80942ddc619
-size 7046931
