+++ package.json
@@ -5,7 +5,8 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "postinstall": "scripts/download.js"
   },
   "repository": "https://github.com/sunrise-choir/jsbot-patchql",
   "keywords": [],
